Eliminar usuarios

+++ controllers/usuarios.controller.js
@@ -1,4 +1,4 @@
-const { usuario, rol,compra, Sequelize } = require('../models')
+const { usuario, rol, compra, Sequelize } = require('../models')
 const jwtSecret = process.env.JWT_SECRET
 const jwt = require('jsonwebtoken')
 const bcrypt = require('bcrypt')
@@ -21,7 +21,7 @@ self.usuarioPutValidator = [
 ];
 
 self.usuarioIDValidator = [
-    param('email').isEmail().withMessage('Debe ser un correo electrónico válido') ,  
+    param('email').isEmail().withMessage('Debe ser un correo electrónico válido'),
 ];
 
 
@@ -32,6 +32,7 @@ self.getAll = async function (req, res, next) {
             attributes: ['id', 'email', 'nombre', [Sequelize.col('rol.nombre'), 'rol']],
             include: { model: rol, attributes: [] }
         })
+        req.bitacora("usuarios.getAll", data.id)
         res.status(200).json(data)
     } catch (error) {
         next(error)
@@ -50,22 +51,22 @@ self.get = async function (req, res, next) {
         if (data) {
             return res.status(200).json(data)
         }
+        req.bitacora("usuarios.get", data.id)
         res.status(404).send()
     } catch (error) {
         next(error)
     }
 }
-
 self.create = async function (req, res, next) {
     try {
         const passwordRegex = /^(?=.*[A-Z])(?=.*\d)(?=.*[!@#$%^&*(),.?":{}|<>]).{8,}$/;
         const token = req.header('Authorization')?.replace('Bearer ', '');
-        const emailExistente = await usuario.findOne({ where: { email: req.body.email } });
-        if (emailExistente) {
-            return res.status(409).send();
-        }
+        const emailExistente = await usuario.findOne({ where: { email: req.body.email } })
         if (!validator.isEmail(req.body.email) || !passwordRegex.test(req.body.password) || !req.body.nombre?.trim()) {
             return res.status(422).send();
+        }
+        if (emailExistente) {
+            return res.status(409).send();
         }
         if (token.startsWith("Bearer")) {
             const rolusuario = await rol.findOne({ where: { nombre: "Usuario" } })
@@ -80,6 +81,7 @@ self.create = async function (req, res, next) {
                 rolid: rolusuario.id,
                 protegido: 0
             })
+            req.bitacora("usuarios.create", data.id)
             res.status(201).json({
                 id: data.id,
                 email: data.email,
@@ -116,6 +118,7 @@ self.create = async function (req, res, next) {
                 rolid: rolusuario.id,
                 protegido: 0
             })
+            req.bitacora("usuarios.create", data.id)
             res.status(201).json({
                 id: data.id,
                 email: data.email,
@@ -148,11 +151,12 @@ self.update = async function (req, res, next) {
         updateData.nombre = req.body.nombre;
         const data = await usuario.update(updateData, {
             where: { id: usuarioRecuperado.id },
-            fields: ['nombre'] 
+            fields: ['nombre']
         });
         if (data[0] === 0) {
             return res.status(404).send()
         }
+        req.bitacora("usuarios.create", data.id)
         res.status(204).send()
     } catch (error) {
         next(error)
@@ -160,33 +164,36 @@ self.update = async function (req, res, next) {
 }
 
 self.delete = async function (req, res, next) {
-    try {
-        //Comprobar si tiene compras, si tiene compras, entonces conflicto
+    try {            
         const errors = validationResult(req)
-        if (!errors.isEmpty()) throw new Error(JSON.stringify(errors))
+        if (!errors.isEmpty()) return res.status(400).send(JSON.stringify(errors));
         const email = req.params.email
-    
+        const token = req.header('Authorization')?.replace('Bearer ', '');
+        const decodedToken = jwt.verify(token, jwtSecret)
+        if (decodedToken == null || decodedToken[ClaimTypes.Name] == null) {
+            return res.status(404).send();
+        }
         let usuarioRecuperado = await usuario.findOne({ where: { email: email } })
         if (usuarioRecuperado == null) {
-            return res.status(404).send()
+            return res.status(404).send();
         }
-        if (usuarioRecuperado.protegido == 1) {
-            return res.status(403).send()
+        if (usuarioRecuperado.protegido == 1 || usuarioRecuperado.email == decodedToken[ClaimTypes.Name]) {
+            return res.status(403).send();
         }
         let comprasCliente = await compra.findAll({
             attributes: ['id'],
             where: { usuarioid: usuarioRecuperado.id }
         });
         if (comprasCliente.length > 0) {
-            return res.status(409).send()
+            return res.status(409).send();
         }
         data = await usuario.destroy({ where: { email: email } })
 
         if (data === 1) {
-            //req.bitacora("usuarios.eliminar", email)
-            return res.status(204).send()
+            return res.status(204).send();
         }
-        res.status(400).send()
+        req.bitacora("usuarios.delete", data)
+        return res.status(400).send();
     } catch (error) {
         next(error)
     }
